requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2
django-cleanup==5.2.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
